requirements.txt
streamlit>=1.30.0
python-bcb>=0.2.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
numpy>=1.24.0
requests>=2.31.0
pyarrow
